requirements.txt
beautifulsoup4==4.12.3
black==24.4.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
docker-pycreds==0.4.0
filelock==3.14.0
fsspec==2024.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.3
idna==3.7
Jinja2==3.1.4
logzero==1.7.0
MarkupSafe==2.1.5
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
protobuf==4.25.3
psutil==5.9.8
pyclean==3.0.0
pycparser==2.22
pyheif==0.7.1
PySocks==1.7.1
PyYAML==6.0.1
requests==2.32.3
safetensors==0.4.3
sentry-sdk==2.4.0
setproctitle==1.3.3
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sympy==1.12.1
timm==1.0.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
wandb==0.17.0
wheel==0.43.0
